requirements.txt
gradio==5.16.0
python-dotenv==1.0.1
ollama==0.4.7
requests==2.32.3
openai==1.63.0
